+++ extensions/conflictSearch/browser/index.js
@@ -58,18 +58,13 @@ var db = urlparser.db;
 /**
  * @type {*|exports|module.exports}
  */
-var noUiSlider = require('nouislider');
-
-/**
- * @type {*|exports|module.exports}
- */
 var io = require('socket.io-client');
 
 /**
  *
  * @type {Element}
  */
-var bufferSlider;
+var sliderEl;
 
 /**
  *
@@ -340,36 +335,29 @@ module.exports = module.exports = {
             me.makeSearch($("#conflict-custom-search").val());
         }, id, false, getProperty);
 
-        bufferSlider = document.getElementById('conflict-buffer-slider');
+
+        sliderEl = $('#conflict-buffer-slider');
         bufferValue = document.getElementById('conflict-buffer-value');
-        try {
-            noUiSlider.create(bufferSlider, {
-                start: startBuffer,
-                connect: "lower",
-                step: 0.01,
-                range: {
-                    min: -5,
-                    max: 500
-                }
-            });
-            bufferSlider.noUiSlider.on('update', function (values, handle) {
-                bufferValue.value = values[handle];
-                currentBufferValue = bufferValue.value;
-            });
-            bufferSlider.noUiSlider.on('change', debounce(function (values, handle) {
-                //currentBufferValue = values[handle];
+        bufferValue.value = startBuffer;
+
+        sliderEl.append(`<div class="range"">
+                                            <input type="range"  min="-5" max="500" value="${startBuffer}" class="js-info-buffer-slider form-range">
+                                            </div>`);
+        let slider = sliderEl.find('.js-info-buffer-slider');
+        slider.on('input change', debounce(function (values) {
+            bufferValue.value = parseFloat(values.target.value);
+            currentBufferValue = bufferValue.value;
+            if (typeof bufferItems._layers[Object.keys(bufferItems._layers)[0]] !== "undefined" && typeof bufferItems._layers[Object.keys(bufferItems._layers)[0]]._leaflet_id !== "undefined") {
                 bufferItems.clearLayers();
                 me.makeSearch()
+            }
+        }, 300));
+        // When the input changes, set the slider value
+        bufferValue.addEventListener('change', function () {
+            slider.val(this.value);
+            slider.trigger('change');
+        });
 
-            }, 300));
-            // When the input changes, set the slider value
-            bufferValue.addEventListener('change', function () {
-                bufferSlider.noUiSlider.set([this.value]);
-            });
-        } catch (e) {
-        }
-
-        // TODO extensios are are initiated AFTER "ready:meta", so below is newer reached
         backboneEvents.get().on("ready:meta", function () {
             metaData = meta.getMetaData();
         })
@@ -385,7 +373,7 @@ module.exports = module.exports = {
         noHitsTable = $("#nohits-content tbody");
         errorTable = $("#error-content tbody");
         let c = 0;
-        backboneEvents.get().on("end:conflictSearch", ()=>{
+        backboneEvents.get().on("end:conflictSearch", () => {
             c = 0;
         })
         // Start listen to the web socket
@@ -511,12 +499,20 @@ module.exports = module.exports = {
         }, 2500);
 
         if (urlparser.urlVars?.var_landsejerlavskode && urlparser.urlVars?.var_matrikelnr) {
-            setTimeout(()=> {
+            setTimeout(() => {
                 if (!fromVarsIsDone) {
                     let placeStore = getPlaceStore();
                     placeStore.db = search.getMDB();
                     placeStore.host = search.getMHOST();
-                    placeStore.sql = `SELECT sfe_ejendomsnummer,ST_Multi(ST_Union(the_geom)),ST_asgeojson(ST_transform(ST_Multi(ST_Union(the_geom)),4326)) as geojson FROM matrikel.jordstykke WHERE sfe_ejendomsnummer = (SELECT sfe_ejendomsnummer FROM matrikel.jordstykke WHERE landsejerlavskode=${urlparser.urlVars.var_landsejerlavskode} AND matrikelnummer='${urlparser.urlVars.var_matrikelnr.toLowerCase()}') group by sfe_ejendomsnummer`;
+                    placeStore.sql = `SELECT sfe_ejendomsnummer,
+                                             ST_Multi(ST_Union(the_geom)),
+                                             ST_asgeojson(ST_transform(ST_Multi(ST_Union(the_geom)), 4326)) as geojson
+                                      FROM matrikel.jordstykke
+                                      WHERE sfe_ejendomsnummer = (SELECT sfe_ejendomsnummer
+                                                                  FROM matrikel.jordstykke
+                                                                  WHERE landsejerlavskode = ${urlparser.urlVars.var_landsejerlavskode}
+                                                                    AND matrikelnummer = '${urlparser.urlVars.var_matrikelnr.toLowerCase()}')
+                                      group by sfe_ejendomsnummer`;
                     placeStore.load();
                     fromVarsIsDone = true;
                 }
@@ -823,8 +819,7 @@ module.exports = module.exports = {
         $('#conflict-main-tabs a[href="#conflict-result-content"]').tab('show');
         if (window.vidiConfig.template === "conflict.tmpl") {
             $('#conflict-result-content a[href="#hits-data-content"]').tab('show');
-        }
-        else {
+        } else {
             $('#conflict-result-content a[href="#hits-content"]').tab('show');
         }
         $('#conflict-open-pdf').attr("href", "/html?id=" + response.file)
@@ -998,8 +993,10 @@ module.exports = module.exports = {
     getBufferItems: function () {
         return bufferItems;
     },
-    setValueForNoUiSlider: function (v) {
-        bufferSlider.noUiSlider.set([v]);
+    setValueForSlider: function (v) {
+        let slider = sliderEl.find('.js-info-buffer-slider');
+        slider.val(v);
+        bufferValue.value = v;
     },
     getFromVarsIsDone: function () {
         return fromVarsIsDone;
@@ -1012,7 +1009,7 @@ let dom = `
         <div>
             <label for="conflict-buffer-value" class="control-label">Buffer</label>
             <input id="conflict-buffer-value" class="form-control">
-            <div id="conflict-buffer-slider" class="slider shor"></div>
+            <div id="conflict-buffer-slider" style="margin-bottom: 20px"></div>
         </div>
     </div>
     <div id="conflict-places" class="places" style="margin-bottom: 20px; display: none">