+++ CHANGELOG.md
@@ -6,9 +6,10 @@ and this project adheres to [CalVer](https://calver.org/).
 
 ## [UNRELEASED] - 2022-17-11
 ### Changed
-- Changed from noUiSlider to native HTML range input:
+- Changed from noUiSlider to native HTML range input in these modules. noUiSlider packages is no longer installed:
   - Base layers
   - Advanced info
+  - Conflict
 
 ## [2022.11.1] - 2022-16-11
 ### Changed